requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
Flask==3.1.3
Flask-Limiter==4.1.1
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
httpcore==1.0.9
httpx==0.28.1
Jinja2==3.1.6
limits==5.8.0
Pillow==12.1.1
pluggy==1.6.0
postgrest==2.28.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pytest==9.0.3
pytest-asyncio==0.26.0
python-dotenv==1.2.2
realtime==2.28.0
requests==2.32.5
sentry-sdk[flask]==2.25.1
structlog==25.3.0
supabase==2.28.0
tenacity==9.1.4
Werkzeug==3.1.6
